# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Golang Quickstart using Gin Gonic",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/route/{id}": {
            "get": {
                "description": "Get Route with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Route collection"
                ],
                "summary": "Get Route Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route ID like route_10000",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Route"
                        }
                    },
                    "404": {
                        "description": "Route Document ID Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "put": {
                "description": "Update Route with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Route collection"
                ],
                "summary": "Update Route Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route ID like route_10000",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updates document",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Route"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Route"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "post": {
                "description": "Create Route with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Route collection"
                ],
                "summary": "Insert Route Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route ID like route_10000",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data to create a document",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Route"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Route"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "409": {
                        "description": "Route Document already exists"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            },
            "delete": {
                "description": "Delete Route with specified ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Route collection"
                ],
                "summary": "Delete Route Document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Route ID like route_10000",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Route Deleted"
                    },
                    "404": {
                        "description": "Route Document ID Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Route": {
            "type": "object",
            "required": [
                "airline",
                "airline_id",
                "destinationairport",
                "sourceairport"
            ],
            "properties": {
                "airline": {
                    "type": "string",
                    "example": "AF"
                },
                "airline_id": {
                    "type": "string",
                    "example": "airline_10"
                },
                "destinationairport": {
                    "type": "string",
                    "example": "JFK"
                },
                "distance": {
                    "type": "number",
                    "example": 4151.79
                },
                "equipment": {
                    "type": "string",
                    "example": "CRJ"
                },
                "schedule": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Schedule"
                    }
                },
                "sourceairport": {
                    "type": "string",
                    "example": "SFO"
                },
                "stops": {
                    "type": "integer",
                    "example": 0
                }
            }
        },
        "models.Schedule": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "integer",
                    "example": 1
                },
                "flight": {
                    "type": "string",
                    "example": "XYZ123"
                },
                "utc": {
                    "type": "string",
                    "example": "14:30"
                }
            }
        }
    }
}